Update Ido_Ohayon
Hi, I just wanted to apologize for the name that appeared by mistake. I’ve been working on this assignment for a long time, making a lot of changes and edits, and honestly I’m feeling pretty frustrated with it at this point. Most of the recommendations I created on my own through comparison and reasoning, but I admit that in some cases I took inspiration from existing ideas because I simply don’t know such a large pool of movies.